Show information about playing track in the UI.
Move some code around so that reading the tags is
not done in the activity thread; all media I/O is
done by the control service thread. The main activity
updates things when events are dispatched by the
service.

Changed some code to save / load state in a more
generic manner.