Don't try to execute actions if SD is not available.
Two part change: first is that the code shouldn't try
to execute actions that touch the SD if the SD is not
available.

The second is that it seems that eject/mount events are
not properly sent by Android 2.2: the app is killed
before any notifications arrive. So the code can't rely
on the notifications to do the right thing, so instead
it checks the state the OS export - which is racy, but
seems to work.