Use MediaPlayer.setNextMediaPlayer().
Try to make transitions less abrupt, something that has
been happening on my phone. Should probably also extend
this to transitions to different albums, but I need to
clean up this code before doing that.

There's probably a narrow race while setting up the next
track with a really slow phone and a short current track,
but that shouldn't be a problem in reality.